requirements.txt
rnet
beautifulsoup4
colorlog
lxml
plotly
streamlit
openai
tqdm
pandas
numpy
python-dotenv 
pymysql